feat: Phase 6 - ストーリー画像をプレースホルダーから本番アセットに差し替え
Canvas APIによるプレースホルダー画像生成を削除し、画像生成AIで作成した
9枚の本番WebP画像をimportする方式に変更。StoryImageEntry型をtypes.tsに移動。
