+++ .docs/20260219-01/tasks.md
@@ -191,9 +191,9 @@
   - 画像キーの一致確認（story.ts ↔ storyImages.ts）
   - 既存テストの更新と実行
 - **受入条件**:
-  - [ ] テキストに文字化け・誤字がない
-  - [ ] 全画像キーが画像レジストリに登録されている
-  - [ ] テスト全パス
+  - [x] テキストに文字化け・誤字がない
+  - [x] 全画像キーが画像レジストリに登録されている
+  - [x] テスト全パス
 
 ### V-2.2: 画面遷移の統合検証
 
@@ -205,9 +205,9 @@
   - ゲームオーバー画面の表示テスト
   - モバイルレイアウトの確認
 - **受入条件**:
-  - [ ] 全画面遷移がエラーなく完了する
-  - [ ] スキップ機能が全画面で動作する
-  - [ ] モバイル表示でレイアウトが崩れない
+  - [x] 全画面遷移がエラーなく完了する
+  - [x] スキップ機能が全画面で動作する
+  - [x] モバイル表示でレイアウトが崩れない
 
 ---
 
@@ -538,10 +538,10 @@
   - 画像内にテキストを含めない
   - 各画像の仕様は `images.md` を参照
 - **受入条件**:
-  - [ ] 9 枚の WebP 画像が生成されている
-  - [ ] 各画像が 960×540 px である
-  - [ ] ファイルサイズが 50〜150 KB の範囲内
-  - [ ] 画像内にテキストが含まれていない
+  - [x] 9 枚の WebP 画像が生成されている
+  - [x] 各画像が 960×540 px である
+  - [x] ファイルサイズが 50〜150 KB の範囲内
+  - [x] 画像内にテキストが含まれていない
 
 ### I-6.2: 画像ファイルの配置
 
@@ -551,7 +551,7 @@
   - 生成した 9 枚の画像を `src/assets/images/` に配置
   - ファイル名規則: `ipne_story_{用途}.webp`
 - **受入条件**:
-  - [ ] 以下の 9 ファイルが配置されている:
+  - [x] 以下の 9 ファイルが配置されている:
     - `ipne_story_prologue_1.webp`
     - `ipne_story_prologue_2.webp`
     - `ipne_story_prologue_3.webp`
@@ -572,10 +572,10 @@
   - `IMAGE_SOURCES` マッピングで画像キー → import を対応付け
   - `getStoryImage()` の返却値を import 済み画像パスに変更
 - **受入条件**:
-  - [ ] Canvas API によるプレースホルダー生成コードが削除されている
-  - [ ] 9 枚の画像が import されている
-  - [ ] `getStoryImage()` が import 済みの画像パスを返す
-  - [ ] `hasStoryImage()` が正しく動作する
+  - [x] Canvas API によるプレースホルダー生成コードが削除されている
+  - [x] 9 枚の画像が import されている
+  - [x] `getStoryImage()` が import 済みの画像パスを返す
+  - [x] `hasStoryImage()` が正しく動作する
 
 ### V-6.1: ビルド検証
 
@@ -585,8 +585,8 @@
   - Webpack が WebP ファイルを正しくバンドルすることを確認
   - TypeScript コンパイルエラーがないことを確認
 - **受入条件**:
-  - [ ] ビルド成功
-  - [ ] コンパイルエラーなし
+  - [x] ビルド成功
+  - [x] コンパイルエラーなし
 
 ### V-6.2: 画面遷移の目視確認
 
@@ -597,10 +597,10 @@
   - 画像のフェードインアニメーションが正しく動作することを確認
   - モバイル表示でレイアウトが崩れないことを確認
 - **受入条件**:
-  - [ ] 全 8 シーン（プロローグ 3 + ステージ 5）で正しい画像が表示される
-  - [ ] 画像が正しいタイミング・シーンで表示される
-  - [ ] フェードインアニメーションが動作する
-  - [ ] モバイル表示で問題なし
+  - [x] 全 8 シーン（プロローグ 3 + ステージ 5）で正しい画像が表示される
+  - [x] 画像が正しいタイミング・シーンで表示される
+  - [x] フェードインアニメーションが動作する
+  - [x] モバイル表示で問題なし
 
 ### V-6.3: 変更不要箇所の確認
 
@@ -611,7 +611,7 @@
   - `FinalClear.tsx` — 修正不要（`getEndingImage()` は別系統）
   - `story.ts` — 修正不要（`imageKey` 文字列のみ）
 - **受入条件**:
-  - [ ] `storyImages.ts` 以外のファイルに変更が不要であることを確認
+  - [x] `storyImages.ts` 以外のファイルに変更が不要であることを確認
 
 ---
